website/requirements.txt
abstract_singleton==1.0.1
aiocache==0.12.1
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
ansicon==1.89.0
anyio==3.6.2
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
Babel==2.11.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.12.2
bidict==0.22.1
binaryornot==0.4.4
black==23.3.0
blessed==1.19.1
blis==0.7.9
boltons @ file:///croot/boltons_1677628692245/work
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.0.1
click==8.1.3
cmake==3.26.4
colorama==0.4.6
conda==23.3.1
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
confection==0.0.4
contourpy==1.0.7
cookiecutter==2.1.1
coverage==7.2.5
cryptography @ file:///croot/cryptography_1677533068310/work
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dalaipy==2.0.2
databases==0.7.0
decorator==4.4.2
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
Django==4.1.5
django-browser-reload==1.6.0
django-cors-headers==4.0.0
django-picklefield==3.1
django-social-share==2.3.0
django-tailwind==3.4.0
dnspython==2.3.0
docker==6.1.2
docutils==0.17.1
duckduckgo-search==2.9.4
easy-pil==0.2.4
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl
exceptiongroup==1.1.1
fastapi==0.95.1
filelock==3.12.0
flake8==6.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.6.0
fuzzywuzzy==0.18.0
gensim==4.3.1
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.48.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-speech==2.19.0
Google-Images-Search==1.4.6
googleapis-common-protos==1.58.0
googletrans==3.1.0a0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.21.0
httpx==0.13.3
huggingface-hub==0.15.1
hyperframe==5.2.0
identify==2.5.24
idna==2.10
imageio==2.30.0
imageio-ffmpeg==0.4.8
imagesize==1.4.1
indic-nlp-library==0.81
indic-transliteration==2.3.40
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
Jinja2==3.1.2
jinja2-time==0.2.0
jinxed==1.2.0
joblib==1.2.0
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
Levenshtein==0.21.0
lit==16.0.6
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.2
matplotlib==3.6.3
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.3
Morfessor==2.0.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.7.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.2
openapi-python-client==0.13.4
orjson==3.8.10
outcome==1.2.0
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
playsound==1.2.2
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
pre-commit==3.3.1
preshed==3.0.8
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.21.12
psycopg2==2.9.6
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pyfiglet==0.8.post1
pyflakes==3.0.1
Pygments==2.14.0
pymdown-extensions==9.11
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-Levenshtein==0.21.0
python-resize-image==1.1.20
python-slugify==7.0.0
python-socketio==5.8.0
pytube==15.0.0
pytz==2022.7.1
PyYAML==6.0
pyyaml_env_tag==0.1
rapidfuzz==3.0.0
readability-lxml==0.8.1
redis==3.5.3
regex==2022.10.31
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
roman==3.3
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.3.1
scipy==1.10.1
searchtweets==1.7.6
selenium==4.1.4
shellingham==1.5.0.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
snscrape==0.6.2.20230320
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.5
Sphinx==5.3.0
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.48
sqlparse==0.4.3
srsly==2.4.6
starlette==0.26.1
sympy==1.12
termcolor==1.1.0
text-unidecode==1.3
thinc==8.1.10
tiktoken==0.3.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==2.0.1
tqdm==4.64.1
transformers==4.30.2
trio==0.22.0
trio-websocket==0.10.2
triton==2.0.0
tweepy==4.14.0
tweet-parser==1.13.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
urllib3-secure-extra==0.1.0
uvicorn==0.22.0
vcrpy==4.2.1
virtualenv==20.23.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webdriver-manager==3.8.6
websocket-client==1.5.1
whitenoise==6.4.0
wordcloud==1.8.2.2
wrapt==1.15.0
wsproto==1.2.0
xx-ent-wiki-sm @ https://github.com/explosion/spacy-models/releases/download/xx_ent_wiki_sm-3.5.0/xx_ent_wiki_sm-3.5.0-py3-none-any.whl
yarl==1.9.2
zstandard @ file:///croot/zstandard_1677013143055/work
